backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
deepface
google-genai
firebase-admin
tf-keras
